+++ rns-api/rns-admin/src/main/java/tech/remote/admin/module/business/goods/domain/form/SalesItemQueryForm.java
@@ -19,12 +19,12 @@ public class SalesItemQueryForm extends PageParam{
     private Long salesId;
 
     @Schema(description = "型号规格ID")
-    private Integer skuId;
+    private Long skuId;
 
     @Schema(description = "商品ID")
-    private Integer goodsId;
+    private Long goodsId;
 
     @Schema(description = "商品类目ID")
-    private Integer categoryId;
+    private Long categoryId;
 
 }
