+++ rns-api/rns-admin/src/main/java/tech/remote/admin/module/business/goods/domain/vo/SalesItemVO.java
@@ -23,13 +23,13 @@ public class SalesItemVO {
     private Long salesId;
 
     @Schema(description = "型号规格ID")
-    private Integer skuId;
+    private Long skuId;
 
     @Schema(description = "商品ID")
-    private Integer goodsId;
+    private Long goodsId;
 
     @Schema(description = "商品类目ID")
-    private Integer categoryId;
+    private Long categoryId;
 
     @Schema(description = "销售数量")
     private Integer quantity;
